Fixed Issues
Attached Square to Mouse Position

+++ Assignment Solution/Assignment Solution/ShapeRendering.cpp
@@ -1,7 +1,7 @@
 #include "ShapeRendering.h"
 namespace ShapeRendering
 {
-	SDL_Rect ShapeRendering::Rectangle(int x, int y, int w, int h)
+	SDL_Rect ShapeRendering::Rectangle(double x, double y, double w, double h)
 	{
 		// Creates the rectangle
 		SDL_Rect r;
